moved memory manager to angr_ext and added a exploration technique called rollback buffer
