+++ src/dAngr/cli/debugger_commands/execution.py
@@ -245,7 +245,7 @@ class ExecutionCommands(BaseCommand):
         """
         return self.debugger.get_from_state(name)
 
-    def load(self, binary_path:str, base_addr:int=0, veritesting:bool=False, history_cache_state:int=0, **kwargs):
+    def load(self, binary_path:str, base_addr:int=0, veritesting:bool=False, **kwargs):
         """
         Load a binary into the debugger.
 
@@ -254,7 +254,6 @@ class ExecutionCommands(BaseCommand):
             base_addr (int): The base address of the binary. Default is 0, means the binary is loaded at its default base address.
             veritesting (bool): Enable veritesting. Default is False.
             kwargs (dict): Additional keyword arguments to pass to the angr project.
-            history_cache_state (int): Tracks the most recent state changes. Default: 0.
 
         Short name: l
         """
@@ -265,7 +264,6 @@ class ExecutionCommands(BaseCommand):
                 raise DebuggerCommandError(f"File '{binary_path}' not found.")
             if base_addr:
                 kwargs['base_addr'] = base_addr
-            kwargs['history_cache_state'] = history_cache_state
             
             self.debugger.init(binary_path, **kwargs)
         except Exception as e:
@@ -452,4 +450,16 @@ class ExecutionCommands(BaseCommand):
 
         self.debugger.undo_step(index)
 
+    def set_exploration_technique(self, technique:str, **kwargs):
+        """
+        Set the exploration technique.
+
+        Args:
+            technique (str): The exploration technique to set.
+            kwargs (dict): Additional keyword arguments to pass to the exploration technique.
+
+        Short name: set
+        """
+        self.debugger.set_exploration_technique(technique, **kwargs)
+        self.send_info(f"Exploration technique set to {technique}.")
